Pass error to next middleware in router
I'm not sure if this is intentional but I had to do this for my tests to pass when porting from `express`. otherwise I get a 404 instead of a 500 when calling `next(error)` in my error handler. we do this in case error handling itself fails to fallback to `express` default error handling.